DTI_MLCD_requirements.txt
arff==0.9
backports.weakref==1.0rc1
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.49.0
boto3==1.10.46
botocore==1.13.46
cairocffi===file-.cairocffi-VERSION
certifi==2016.2.28
cffi==1.14.0
Click==7.0
community==1.0.0b1
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
entrypoints==0.2.3
Flask==1.1.1
future==0.16.0
gensim==3.6.0
html5lib==0.9999999
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
iterative-stratification==0.1.6
itsdangerous==1.1.0
jedi==0.10.2
jieba==0.39
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Keras==2.0.5
kiwisolver==1.1.0
liac-arff==2.2.1
lightgbm==2.2.3
lxml==4.4.2
Mako==1.0.6
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==3.1.2
mistune==0.7.4
mlxtend==0.17.0
MolVS==0.1.1
nbconvert==5.2.1
nbformat==4.4.0
networkx==2.3
nose==1.3.7
notebook==5.0.0
numpy==1.17.3
olefile==0.44
pandas==0.25.3
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pycairo==1.19.1
pycparser==2.19
Pygments==2.2.0
pygpu==0.6.9
pyparsing==2.2.0
python-dateutil==2.6.1
python-igraph==0.8.0
python-louvain==0.13
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
requests==2.14.2
s3transfer==0.2.1
scikit-learn==0.22
scikit-multilearn==0.2.0
scipy==1.3.2
seaborn==0.9.0
simplegeneric==0.8.1
six==1.10.0
smart-open==1.9.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.5
terminado==0.6
testpath==0.3
texttable==1.6.2
Theano==0.9.0
tornado==4.5.2
tqdm==4.15.0
traitlets==4.3.2
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
widgetsnbextension==3.0.2
xgboost==0.90
xmind==0.1a0
